requirements.txt
Django==1.11.6
django-extensions==1.9.3
djangorestframework==3.7.1
django-model-utils==3.0.0
django-filter==1.0.4
factory-boy==2.9.2
MySQL-python==1.2.5
PyMySQL==0.7.9
